create JPEG class
